Here Are my Edits to the group sorter project.
Here are my edits for the groups-sorter app, along with its other dependencies such as the extra icons and added css and javascript. I created a new UI design for the sorter application. So, now, the user can click on a subject and it'll bring the time groups up on separate columns. One column for Full Time, and the other for Part Time. Please, take a look at the differences and review. Also, feel free to add or take what you may to make this a better version. :)